Add files via upload
If you are a teenager or the parent of a teenagers, this self help book is for you. How do you know what to do and how to behave as an adult? What is importnat in life and why? Who am I and where am I going? What are my chances of finding the right partner and how do I choose a partner? Find out exactly what, how and when to do to become a successfull adult with a ourpose in life and living your best life ever, being your best self ever. Get going today.